Moved functions and added more documentation
