+++ vision/pipeline/standard_object.py
@@ -1,26 +1,23 @@
+"""Functions that perform standard object detection, localization, and classification"""
+
+from typing import TypeAlias
+
 import cv2
 import numpy as np
-import json
 
 from nptyping import NDArray, Shape, UInt8
-from typing import Callable, TypeAlias
 import vision.common.constants as consts
 
 from vision.competition_inputs.bottle_reader import BottleData
 from vision.common.bounding_box import BoundingBox
 from vision.common.odlc_characteristics import ODLCColor
 
-# from vision.emergent_object.emergent_object import detect_emergent_object
-
 from vision.standard_object.odlc_image_processing import preprocess_std_odlc
 from vision.standard_object.odlc_classify_shape import process_shapes
 from vision.standard_object.odlc_text_detection import get_odlc_text
 from vision.standard_object.odlc_colors import find_colors
 
-from vision.deskew.camera_distances import get_coordinates
-
-from vision.pipeline.pipeline_utils import *
-
+import vision.pipeline.pipeline_utils as utils
 
 ContourHeirarchyList: TypeAlias = list[tuple[tuple[consts.Contour, ...], consts.Hierarchy]]
 
@@ -34,8 +31,8 @@ PROCESSING_THRESHOLDS: list[tuple[int, int]] = [
 
 def find_standard_objects(
     original_image: consts.Image,
-    # camera_parameters: consts.CameraParameters,
-    # image_path: str
+    camera_parameters: consts.CameraParameters,
+    image_path: str
 ) -> list[BoundingBox]:
     """
     Finds all bounding boxes of standard objects in an image
@@ -72,8 +69,10 @@ def find_standard_objects(
                 continue  # Skip the current shape and move on to the next
 
             # Modifies by reference
-            # if not set_generic_attributes(shape, image_path, original_image.shape, camera_parameters):
-            #     continue
+            if not utils.set_generic_attributes(
+                shape, image_path, original_image.shape, camera_parameters
+            ):
+                continue
 
             found_odlcs.append(shape)
 
@@ -97,8 +96,8 @@ def iterate_find_contours(original_image: consts.Image) -> ContourHeirarchyList:
 
     contour_heirarchies_list: ContourHeirarchyList = []
 
-    for range in PROCESSING_THRESHOLDS:
-        processed_image = preprocess_std_odlc(original_image, range[0], range[1])
+    for thresholds in PROCESSING_THRESHOLDS:
+        processed_image = preprocess_std_odlc(original_image, thresholds[0], thresholds[1])
 
         # Get the contours in the image
         contours: tuple[consts.Contour, ...]
@@ -110,37 +109,6 @@ def iterate_find_contours(original_image: consts.Image) -> ContourHeirarchyList:
         contour_heirarchies_list.append((contours, hierarchy))
 
     return contour_heirarchies_list
-
-
-def sort_odlcs(
-    bottle_info: list[BottleData], saved_odlcs: list[BoundingBox]
-) -> list[list[BoundingBox]]:
-    """
-    Sorts the standard objects in the given list by which bottle they match
-
-    Parameters
-    ----------
-    bottle_info: list[BottleData]
-        The data describing the object matching each bottle
-    saved_odlcs: list[BoundingBox]
-        The list of all sightings of standard objects
-    Returns
-    -------
-    sorted_odlcs: list[list[BoundingBox]]
-        The list of sightings of each object, matched to bottles
-    """
-
-    # The first index represents the bottle index - that's why there's 5
-    sorted_odlcs: list[list[BoundingBox]] = [[], [], [], [], []]
-
-    for shape in saved_odlcs:
-        bottle_index: int = get_bottle_index(shape, bottle_info)
-
-        # Save the shape bounding box in its proper place
-        if bottle_index != -1:
-            sorted_odlcs[bottle_index].append(shape)
-
-    return sorted_odlcs
 
 
 def set_shape_attributes(
@@ -185,6 +153,37 @@ def set_shape_attributes(
     shape.set_attribute("text_color", text_color)
 
     return True
+
+
+def sort_odlcs(
+    bottle_info: list[BottleData], saved_odlcs: list[BoundingBox]
+) -> list[list[BoundingBox]]:
+    """
+    Sorts the standard objects in the given list by which bottle they match
+
+    Parameters
+    ----------
+    bottle_info: list[BottleData]
+        The data describing the object matching each bottle
+    saved_odlcs: list[BoundingBox]
+        The list of all sightings of standard objects
+    Returns
+    -------
+    sorted_odlcs: list[list[BoundingBox]]
+        The list of sightings of each object, matched to bottles
+    """
+
+    # The first index represents the bottle index - that's why there's 5
+    sorted_odlcs: list[list[BoundingBox]] = [[], [], [], [], []]
+
+    for shape in saved_odlcs:
+        bottle_index: int = get_bottle_index(shape, bottle_info)
+
+        # Save the shape bounding box in its proper place
+        if bottle_index != -1:
+            sorted_odlcs[bottle_index].append(shape)
+
+    return sorted_odlcs
 
 
 def get_bottle_index(shape: BoundingBox, bottle_info: list[BottleData]) -> int:
@@ -234,5 +233,37 @@ def get_bottle_index(shape: BoundingBox, bottle_info: list[BottleData]) -> int:
         # Gets the index of the first bottle with the most matches.
         # First [0] takes the first dimension, second [0] takes the first element
         return np.where(all_matches == all_matches.max())[0][0]
-    else:
-        return -1
+
+    return -1
+
+
+def create_odlc_dict(sorted_odlcs: list[list[BoundingBox]]) -> consts.ODLC_Dict:
+    """
+    Creates the ODLC_Dict dictionary from a list of shape bounding boxes
+
+    Parameters
+    ----------
+    list[list[BoundingBox]]
+        The list of sightings of each object, matched to bottles
+
+    Returns
+    -------
+    odlc_dict: consts.ODLC_Dict
+        The dictionary of ODLCs matching the output format
+    """
+
+    odlc_dict: consts.ODLC_Dict = {}
+
+    for i, bottle in enumerate(sorted_odlcs):
+        coords_list: list[tuple[int, int]] = []
+
+        for shape in bottle:
+            coords_list.append((shape.get_attribute("latitude"), shape.get_attribute("longitude")))
+
+        coords_array = np.array(coords_list)
+
+        average_coord = np.average(coords_array, axis=0)
+
+        odlc_dict[str(i)] = {"latitude": average_coord[0], "longitude": average_coord[1]}
+
+    return odlc_dict
